requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
environs==10.0.0
Pillow~=9.5.0
